Made line and col a single type
 - Lots of minor refactoring
 - A File is now required to parse

+++ include/lexer.hpp
@@ -19,7 +19,7 @@ std::string next (const char * source, size_t & pos, size_t length);
 tkstring lex (const char * source, size_t length);
 inline tkstring lex (const std::string & source) {return lex (source.c_str (), source.length ());}
 inline tkstring lexFile (const std::string & source) {return lex (source.c_str (), source.length ()) + token (TOKEN::eof, "End of file");}
-tkstring lexline (const char * source, size_t & pos, size_t length, size_t line=0);
+tkstring lexline (const char * source, size_t & pos, size_t length, uint32_t line=0);
 token lexToken (const std::string & source);
